move RocksDB to crate::store  and create BlockStore trait
