+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod block;
 pub mod client;
 pub mod server;
 pub mod state;
+pub mod store;
 
 pub const SOCKET_NAME: &str = "@mina-indexer.sock";
 pub const MAINNET_TRANSITION_FRONTIER_K: u32 = 290;
