+++ tests/block/store/add_and_get_blocks.rs
@@ -1,6 +1,6 @@
 use std::{collections::HashMap, path::PathBuf};
 
-use mina_indexer::block::{parser::BlockParser, store::BlockStoreConn};
+use mina_indexer::{block::{parser::BlockParser, store::BlockStore}, store::IndexerStore};
 use tokio::time::Instant;
 
 #[tokio::test]
@@ -8,7 +8,7 @@ async fn rocksdb() {
     let store_dir = &PathBuf::from("./block-store-test");
     let log_dir = &PathBuf::from("./tests/data/sequential_blocks");
 
-    let db = BlockStoreConn::new(store_dir).unwrap();
+    let db = IndexerStore::new(store_dir).unwrap();
     let mut bp = BlockParser::new(log_dir).unwrap();
     let mut blocks = HashMap::new();
